+++ src/elements/dialog/dialog.component.ts
@@ -1,0 +1,13 @@
+import { Component, Inject } from '@angular/core';
+import {MAT_DIALOG_DATA} from '@angular/material/dialog';
+
+import { Superhero } from '../../interfaces/superhero.interface';
+
+@Component({
+  selector: 'app-dialog',
+  templateUrl: './dialog.component.html',
+  styleUrls: ['./dialog.component.scss']
+})
+export class DialogComponent {
+  constructor(@Inject(MAT_DIALOG_DATA) public data: Superhero) { }
+}