+++ src/elements/table/table.component.ts
@@ -1,4 +1,4 @@
-import { AfterViewInit, Component, Input, ViewChild } from '@angular/core';
+import { AfterViewInit, Component, EventEmitter, Input, Output, ViewChild } from '@angular/core';
 import { MatSort } from '@angular/material/sort';
 import { MatTable, MatTableDataSource } from '@angular/material/table';
 import { Observable } from 'rxjs';
@@ -17,11 +17,16 @@ export class TableComponent implements AfterViewInit {
   @ViewChild(MatTable) table: MatTable<any>;
   @Input() columnsToDisplay: string[];
   @Input() superheroes: Observable<Superhero[]>;
+  @Output() rowSelected = new EventEmitter<Superhero>();
 
   ngAfterViewInit(): void {
     this.superheroes.subscribe( (data: Superhero[]) => {
       this.superheroesMatTable = new MatTableDataSource(data);
       this.superheroesMatTable.sort = this.sort;
     })
+  }
+
+  public selectSuperhero(superhero: Superhero) {
+    this.rowSelected.emit(superhero);
   }
 }