+++ CHANGELOG.md
@@ -17,3 +17,4 @@
 - Move table to a new component into elements folder
 - Solve issues deploying in gitlab, wrong assets route,...
 - Add chips component into elements folder, first version with functions but without table updating
+- Add dialog component into elements folder to be shown when click on any of the superheroes to show his/her info